+++ .env.development
@@ -1,4 +1,4 @@
-API_URL=https://codebility-be.onrender.com/api/v1/development
+NEXT_PUBLIC_BASE_API=https://codebility-be.onrender.com/api/v1/production
 
 MONGO_URI=mongodb+srv://codebility23:codebility0123@codebility.ffkncak.mongodb.net/codebility?retryWrites=true&w=majority
